# laravel-dashboard/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Scraper API",
        "description": "API for managing companies and job postings scraped from LinkedIn",
        "contact": {
            "email": "admin@jobscraper.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/jobs/exists": {
            "get": {
                "tags": [
                    "Job Postings"
                ],
                "summary": "Check if a job posting exists by LinkedIn job ID",
                "description": "Returns whether a job posting with the given LinkedIn job ID exists in the database",
                "operationId": "0f2104a6163cddfd45deab527b7f5d05",
                "parameters": [
                    {
                        "name": "linkedin_job_id",
                        "in": "query",
                        "description": "LinkedIn job ID to search for",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job posting existence check result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "exists": {
                                            "description": "Whether the job posting exists",
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key for authentication",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
}